# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AircraftResponse:
    properties:
      id:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      seats:
        items:
          $ref: '#/definitions/dto.AircraftSeatResponse'
        type: array
      total_seats:
        type: integer
    type: object
  dto.AircraftSeatResponse:
    properties:
      id:
        type: string
      is_exit_row:
        type: boolean
      row_number:
        type: integer
      seat_class:
        $ref: '#/definitions/dto.SeatClassResponse'
      seat_letter:
        type: string
      seat_number:
        type: string
      seat_type:
        type: string
      x_position:
        type: integer
      y_position:
        type: integer
    type: object
  dto.AirportResponse:
    properties:
      city:
        type: string
      code:
        type: string
      country:
        type: string
      id:
        type: string
      name:
        type: string
      timezone:
        type: string
    type: object
...
  dto.BoardingPassResponse:
    properties:
      boarding_group:
        type: string
      boarding_time:
        type: string
      departure_time:
        type: string
      destination:
        type: string
      flight_number:
        type: string
      gate:
        type: string
      id:
        type: string
      origin:
        type: string
      passenger_id:
        type: string
      passenger_name:
        type: string
      qr_code:
        type: string
      seat_number:
        type: string
      segment_id:
        type: string
    type: object
...
  dto.CheckinRequest:
    properties:
      passenger_id:
        type: string
      segment_id:
        type: string
    required:
    - passenger_id
    - segment_id
    type: object
  dto.CheckinResponse:
    properties:
      checkin_time:
        type: string
      id:
        type: string
      passenger_id:
        type: string
      segment_id:
        type: string
    type: object
  dto.CheckinResultResponse:
    properties:
      boarding_pass:
        $ref: '#/definitions/dto.BoardingPassResponse'
      checkin:
        $ref: '#/definitions/dto.CheckinResponse'
    type: object
...
  dto.FlightListResponse:
    properties:
      flights:
        items:
          $ref: '#/definitions/dto.FlightResponse'
        type: array
    type: object
  dto.FlightResponse:
    properties:
      aircraft:
        $ref: '#/definitions/dto.AircraftResponse'
      arrival_time:
        type: string
      available_seats:
        type: integer
      departure_time:
        type: string
      duration_minutes:
        type: integer
      flight_number:
        type: string
      id:
        type: string
      lowest_price:
        type: number
      schedule:
        $ref: '#/definitions/dto.FlightScheduleResponse'
      status:
        $ref: '#/definitions/models.FlightStatus'
    type: object
  dto.FlightScheduleResponse:
    properties:
      arrival_airport:
        $ref: '#/definitions/dto.AirportResponse'
      arrival_time:
        type: string
      departure_airport:
        $ref: '#/definitions/dto.AirportResponse'
      departure_time:
        type: string
      flight_number:
        type: string
      id:
        type: string
      operating_days:
        type: string
    type: object
  dto.FlightSeatResult:
    properties:
      class_code:
        type: string
      class_name:
        type: string
      flight_seat:
        $ref: '#/definitions/models.FlightSeat'
      is_exit_row:
        type: boolean
      row_number:
        type: integer
      seat_letter:
        type: string
      seat_number:
        type: string
      seat_type:
        type: string
    type: object
...
  dto.SeatClassResponse:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.Aircraft:
    properties:
      id:
        type: string
      manufacturer:
        maxLength: 100
        type: string
      model:
        maxLength: 100
        type: string
      seats:
        items:
          $ref: '#/definitions/models.AircraftSeat'
        type: array
      total_seats:
        minimum: 0
        type: integer
    type: object
  models.AircraftSeat:
    properties:
      aircraft:
        $ref: '#/definitions/models.Aircraft'
      aircraft_id:
        type: string
      id:
        type: string
      is_exit_row:
        type: boolean
      row_number:
        minimum: 1
        type: integer
      seat_class:
        $ref: '#/definitions/models.SeatClass'
      seat_class_id:
        type: string
      seat_letter:
        maxLength: 1
        type: string
      seat_number:
        maxLength: 5
        type: string
      seat_type:
        maxLength: 20
        type: string
      x_position:
        type: integer
      y_position:
        type: integer
    required:
    - aircraft_id
    - row_number
    - seat_number
    type: object
  models.Airport:
    properties:
      city:
        maxLength: 255
        type: string
      code:
        type: string
      country:
        maxLength: 255
        type: string
      id:
        type: string
      name:
        maxLength: 255
        type: string
      timezone:
        maxLength: 100
        type: string
    required:
    - code
    type: object
...
  models.Flight:
    properties:
      aircraft:
        $ref: '#/definitions/models.Aircraft'
      aircraft_id:
        type: string
      arrival_time:
        type: string
      departure_time:
        type: string
      flight_seats:
        items:
          $ref: '#/definitions/models.FlightSeat'
        type: array
      id:
        type: string
      schedule:
        $ref: '#/definitions/models.FlightSchedule'
      schedule_id:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.FlightStatus'
        enum:
        - scheduled
        - boarding
        - departed
        - arrived
        - cancelled
        - delayed
    type: object
  models.FlightSchedule:
    properties:
      arrival_airport:
        $ref: '#/definitions/models.Airport'
      arrival_airport_id:
        type: string
      arrival_time:
        type: string
      departure_airport:
        $ref: '#/definitions/models.Airport'
      departure_airport_id:
        type: string
      departure_time:
        type: string
      flight_number:
        maxLength: 10
        type: string
      flights:
        items:
          $ref: '#/definitions/models.Flight'
        type: array
      id:
        type: string
      operating_days:
        type: string
    type: object
  models.FlightSeat:
    properties:
      aircraft_seat:
        $ref: '#/definitions/models.AircraftSeat'
      aircraft_seat_id:
        type: string
      flight:
        $ref: '#/definitions/models.Flight'
      flight_id:
        type: string
      id:
        type: string
      price:
        minimum: 0
        type: number
      status:
        allOf:
        - $ref: '#/definitions/models.FlightSeatStatus'
        enum:
        - available
        - locked
        - booked
    type: object
  models.FlightSeatStatus:
    enum:
    - available
    - locked
    - booked
    type: string
    x-enum-varnames:
    - FlightSeatAvailable
    - FlightSeatLocked
    - FlightSeatBooked
  models.FlightStatus:
    enum:
    - scheduled
    - boarding
    - departed
    - arrived
    - cancelled
    - delayed
    type: string
    x-enum-varnames:
    - FlightStatusScheduled
    - FlightStatusBoarding
    - FlightStatusDeparted
    - FlightStatusArrived
    - FlightStatusCancelled
    - FlightStatusDelayed
...
  models.SeatClass:
    properties:
      code:
        maxLength: 10
        type: string
      id:
        type: string
      name:
        maxLength: 50
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@airline.com
    name: API Support
  description: REST API untuk sistem Passenger Service System (PSS) maskapai penerbangan.
    Mencakup manajemen penerbangan, pemesanan tiket, check-in, bagasi, dan boarding
    pass.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Passenger Service Backend API
  version: "1.0"
paths:
...
  /checkin:
    post:
      consumes:
      - application/json
      description: Melakukan check-in penumpang untuk segment penerbangan. Check-in
        dibuka 24 jam sebelum dan ditutup 45 menit sebelum keberangkatan. Tiket harus
        sudah diterbitkan.
      parameters:
      - description: Data check-in
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CheckinResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Sudah check-in
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Di luar window check-in atau tiket belum ada
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Lakukan check-in
      tags:
      - CheckIn
  /checkin/passenger/{passengerID}:
    get:
      description: Mengambil seluruh riwayat check-in seorang penumpang.
      parameters:
      - description: Passenger UUID
        in: path
        name: passengerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CheckinResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Riwayat check-in penumpang
      tags:
      - CheckIn
  /flights/{id}:
    get:
      description: Mengambil detail satu penerbangan beserta jadwal dan aircraft.
        Endpoint ini bersifat publik.
      parameters:
      - description: Flight UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FlightResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Detail penerbangan
      tags:
      - Flight
  /flights/{id}/seat-map:
    get:
      description: Mengambil seluruh peta kursi penerbangan beserta status ketersediaan
        dan harga. Membutuhkan login.
      parameters:
      - description: Flight UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FlightSeatResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Peta kursi penerbangan
      tags:
      - Flight
  /flights/{id}/status:
    patch:
      consumes:
      - application/json
      description: Mengubah status penerbangan (scheduled, boarding, departed, arrived,
        cancelled, delayed). Hanya admin.
      parameters:
      - description: Flight UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status baru
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update status penerbangan
      tags:
      - Flight
  /flights/generate:
    post:
      consumes:
      - application/json
      description: Membuat instance penerbangan dari sebuah jadwal untuk rentang tanggal
        tertentu. Hanya admin.
      parameters:
      - description: Parameter generate
        in: body
        name: request
        required: true
        schema:
          properties:
            from:
              type: string
            schedule_id:
              type: string
            to:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    generated:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Generate penerbangan dari jadwal
      tags:
      - Flight
  /flights/search:
    get:
      description: Mencari penerbangan yang tersedia berdasarkan rute dan tanggal.
        Endpoint ini bersifat publik.
      parameters:
      - description: Kode IATA keberangkatan
        example: CGK
        in: query
        name: dep
        required: true
        type: string
      - description: Kode IATA tujuan
        example: DPS
        in: query
        name: arr
        required: true
        type: string
      - description: Tanggal (YYYY-MM-DD)
        example: "2026-05-15"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FlightListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Airport tidak ditemukan
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Cari penerbangan tersedia
      tags:
      - Flight
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"